+++ inst/code_paper/code_py_to_html.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 ## This bash script requires jupytext and nbconvert to be installed. They can be installed using:
-# pip install jupytext nbconvert
+# pip install jupyter jupytext nbconvert
 
 ## The code_py.py script ends by listing session information, provided by the session_info package, which needs
 ## to be installed alongside shaprpy. session_info can be installed by running:
@@ -17,7 +17,7 @@ EXECUTED_NOTEBOOK="code_py_executed.ipynb"
 HTMLFILE="code_py.html"
 
 # Step 0: Check for required packages and provide installation commands if missing
-REQUIRED_PACKAGES=("jupytext" "nbconvert")
+REQUIRED_PACKAGES=("jupyter" "jupytext" "nbconvert")
 for PACKAGE in "${REQUIRED_PACKAGES[@]}"; do
     if ! python3 -m pip show "$PACKAGE" > /dev/null 2>&1; then
         echo "Package '$PACKAGE' is not installed. You can install it by running:"
@@ -35,7 +35,9 @@ jupyter nbconvert --to notebook --execute "$NOTEBOOK" --output "$EXECUTED_NOTEBO
 
 # Step 3: Convert executed notebook to HTML
 echo "Exporting to HTML..."
-jupyter nbconvert --to html "$EXECUTED_NOTEBOOK" --output "$HTMLFILE"
+jupyter nbconvert --to html "$EXECUTED_NOTEBOOK" --output "$HTMLFILE" \
+  --TagRemovePreprocessor.enabled=True \
+  --TagRemovePreprocessor.remove_input_tags="['hide_input']" 2>/dev/null
 
 # Step 4: Cleanup intermediate files
 echo "Cleaning up..."